aletheia-research-agent/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.1
langgraph==0.2.0
langchain==0.1.20
langchain-core==0.1.52
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.6
aiohttp==3.9.1
httpx==0.26.0
bleach==6.1.0
supabase==2.3.0
postgrest==0.13.0
python-dotenv==1.0.0
